Remove D2L.CodeStyle.Analysis, and package everything in a single assembly.
