+++ D2L.CodeStyle.sln
@@ -5,15 +5,11 @@ VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "D2L.CodeStyle.Annotations", "src\D2L.CodeStyle.Annotations\D2L.CodeStyle.Annotations.csproj", "{1812AF7E-59B1-4764-8090-F84446A71C4E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "D2L.CodeStyle.Analysis", "src\D2L.CodeStyle.Analysis\D2L.CodeStyle.Analysis.csproj", "{76C3649D-EE9F-45BD-851C-8657EE1CD415}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Test", "Test", "{3873E5AD-A61F-4CE4-B2E1-5AD83D25BF79}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "D2L.CodeStyle.Analyzers", "src\D2L.CodeStyle.Analyzers\D2L.CodeStyle.Analyzers.csproj", "{FE7608CD-C8B1-43DD-BBCC-9B1917C9CA73}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "D2L.CodeStyle.Analyzers.Tests", "tests\D2L.CodeStyle.Analyzers.Test\D2L.CodeStyle.Analyzers.Tests.csproj", "{1C54E78A-E12B-45C3-BA6E-31DB5C63B389}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "D2L.CodeStyle.Analysis.UnitTests", "tests\D2L.CodeStyle.Analysis.UnitTests\D2L.CodeStyle.Analysis.UnitTests.csproj", "{E5E9D471-BB72-4F77-88F8-3B4F77A7D23B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -25,10 +21,6 @@ Global
 		{1812AF7E-59B1-4764-8090-F84446A71C4E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1812AF7E-59B1-4764-8090-F84446A71C4E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1812AF7E-59B1-4764-8090-F84446A71C4E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{76C3649D-EE9F-45BD-851C-8657EE1CD415}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{76C3649D-EE9F-45BD-851C-8657EE1CD415}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{76C3649D-EE9F-45BD-851C-8657EE1CD415}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{76C3649D-EE9F-45BD-851C-8657EE1CD415}.Release|Any CPU.Build.0 = Release|Any CPU
 		{FE7608CD-C8B1-43DD-BBCC-9B1917C9CA73}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{FE7608CD-C8B1-43DD-BBCC-9B1917C9CA73}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FE7608CD-C8B1-43DD-BBCC-9B1917C9CA73}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -37,16 +29,11 @@ Global
 		{1C54E78A-E12B-45C3-BA6E-31DB5C63B389}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1C54E78A-E12B-45C3-BA6E-31DB5C63B389}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1C54E78A-E12B-45C3-BA6E-31DB5C63B389}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E5E9D471-BB72-4F77-88F8-3B4F77A7D23B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E5E9D471-BB72-4F77-88F8-3B4F77A7D23B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E5E9D471-BB72-4F77-88F8-3B4F77A7D23B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E5E9D471-BB72-4F77-88F8-3B4F77A7D23B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{1C54E78A-E12B-45C3-BA6E-31DB5C63B389} = {3873E5AD-A61F-4CE4-B2E1-5AD83D25BF79}
-		{E5E9D471-BB72-4F77-88F8-3B4F77A7D23B} = {3873E5AD-A61F-4CE4-B2E1-5AD83D25BF79}
 	EndGlobalSection
 EndGlobal
