--- src/D2L.CodeStyle.Analysis/BadStaticReason.cs
@@ -1,6 +1,0 @@
-﻿namespace D2L.CodeStyle.Analysis {
-	public enum BadStaticReason {
-		StaticIsMutable,
-		TypeOfStaticIsMutable
-	}
-}
