--- src/D2L.CodeStyle.Analysis/D2L.CodeStyle.Analysis.nuspec
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
-	<metadata>
-		<id>$id$</id>
-		<version>$version$</version>
-		<title>$title$</title>
-		<authors>$author$</authors>
-		<owners>$author$</owners>
-
-		<projectUrl>https://github.com/Brightspace/D2L.CodeStyle</projectUrl>
-		<requireLicenseAcceptance>false</requireLicenseAcceptance>
-		<description>$description$</description>
-		<copyright>$copyright$</copyright>
-
-		<dependencies>
-			<dependency id="D2L.CodeStyle.Annotations" version="$version$" />
-		</dependencies>
-	</metadata>
-</package>
