--- src/D2L.CodeStyle.Analysis/Properties/AssemblyInfo.cs
@@ -1,18 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("D2L.CodeStyle.Analysis")]
-[assembly: AssemblyDescription("Build integrations")]
-[assembly: AssemblyCompany("D2L")]
-[assembly: AssemblyProduct("D2L.CodeStyle.Analysis")]
-[assembly: AssemblyCopyright("Copyright © D2L 2016")]
-
-[assembly: ComVisible(false)]
-
-[assembly: Guid("76c3649d-ee9f-45bd-851c-8657ee1cd415")]
-
-[assembly: AssemblyVersion("0.4.4.0")]
-[assembly: AssemblyFileVersion("0.4.4.0")]
-
-[assembly: InternalsVisibleTo("D2L.CodeStyle.Analysis.UnitTests")]
