--- src/D2L.CodeStyle.Analysis/packages.config
@@ -1,22 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.CodeAnalysis" version="1.2.0" targetFramework="net461" />
-  <package id="Microsoft.CodeAnalysis.Analyzers" version="1.1.0" targetFramework="net461" />
-  <package id="Microsoft.CodeAnalysis.Common" version="1.2.0" targetFramework="net461" />
-  <package id="Microsoft.CodeAnalysis.CSharp" version="1.2.0" targetFramework="net461" />
-  <package id="Microsoft.CodeAnalysis.CSharp.Workspaces" version="1.2.0" targetFramework="net461" />
-  <package id="Microsoft.CodeAnalysis.VisualBasic" version="1.2.0" targetFramework="net461" />
-  <package id="Microsoft.CodeAnalysis.VisualBasic.Workspaces" version="1.2.0" targetFramework="net461" />
-  <package id="Microsoft.CodeAnalysis.Workspaces.Common" version="1.2.0" targetFramework="net461" />
-  <package id="Microsoft.Composition" version="1.0.27" targetFramework="net461" />
-  <package id="System.Collections" version="4.0.0" targetFramework="net461" />
-  <package id="System.Collections.Immutable" version="1.1.37" targetFramework="net461" />
-  <package id="System.Diagnostics.Debug" version="4.0.0" targetFramework="net461" />
-  <package id="System.Globalization" version="4.0.0" targetFramework="net461" />
-  <package id="System.Linq" version="4.0.0" targetFramework="net461" />
-  <package id="System.Reflection.Metadata" version="1.2.0" targetFramework="net461" />
-  <package id="System.Resources.ResourceManager" version="4.0.0" targetFramework="net461" />
-  <package id="System.Runtime" version="4.0.0" targetFramework="net461" />
-  <package id="System.Runtime.Extensions" version="4.0.0" targetFramework="net461" />
-  <package id="System.Threading" version="4.0.0" targetFramework="net461" />
-</packages>
