+++ src/D2L.CodeStyle.Analyzers/D2L.CodeStyle.Analyzers.csproj
@@ -30,9 +30,12 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="ParallelizableTestsAnalyzer.cs" />
+    <Compile Include="Common\MutabilityInspector.cs" />
+    <Compile Include="Common\TypeSymbolExtensions.cs" />
+    <Compile Include="Common\Utils.cs" />
+    <Compile Include="ParallelizableTests\ParallelizableTestsAnalyzer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="UnsafeStaticsAnalyzer.cs" />
+    <Compile Include="UnsafeStatics\UnsafeStaticsAnalyzer.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="D2L.CodeStyle.Analyzers.nuspec">
@@ -50,16 +53,6 @@
   <ItemGroup>
     <Analyzer Include="..\..\packages\Microsoft.CodeAnalysis.Analyzers.1.1.0\analyzers\dotnet\cs\Microsoft.CodeAnalysis.Analyzers.dll" />
     <Analyzer Include="..\..\packages\Microsoft.CodeAnalysis.Analyzers.1.1.0\analyzers\dotnet\cs\Microsoft.CodeAnalysis.CSharp.Analyzers.dll" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\D2L.CodeStyle.Analysis\D2L.CodeStyle.Analysis.csproj">
-      <Project>{76c3649d-ee9f-45bd-851c-8657ee1cd415}</Project>
-      <Name>D2L.CodeStyle.Analysis</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\D2L.CodeStyle.Annotations\D2L.CodeStyle.Annotations.csproj">
-      <Project>{1812af7e-59b1-4764-8090-f84446a71c4e}</Project>
-      <Name>D2L.CodeStyle.Annotations</Name>
-    </ProjectReference>
   </ItemGroup>
   <ItemGroup>
     <Reference Include="Microsoft.CodeAnalysis, Version=1.2.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
