+++ src/D2L.CodeStyle.Analyzers/D2L.CodeStyle.Analyzers.nuspec
@@ -15,10 +15,6 @@
     <frameworkAssemblies>
       <frameworkAssembly assemblyName="System" targetFramework="" />
     </frameworkAssemblies>
-
-    <dependencies>
-      <dependency id="D2L.CodeStyle.Analysis" version="$version$" />
-    </dependencies>
   </metadata>
   
   <!-- The convention for analyzers is to put language agnostic dlls in analyzers\portable50 and language specific analyzers in either analyzers\portable50\cs or analyzers\portable50\vb -->
