+++ src/D2L.CodeStyle.Analyzers/UnsafeStatics/UnsafeStaticsAnalyzer.cs
@@ -1,0 +1,195 @@
+﻿using System.Collections.Immutable;
+using System.Linq;
+using D2L.CodeStyle.Analyzers.Common;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Diagnostics;
+
+namespace D2L.CodeStyle.Analyzers.UnsafeStatics {
+	[DiagnosticAnalyzer( LanguageNames.CSharp )]
+	public sealed class UnsafeStaticsAnalyzer : DiagnosticAnalyzer {
+
+		public const string DiagnosticId = "D2L0002";
+		private const string Category = "Safety";
+
+		private const string Title = "Ensure that static field is safe in undifferentiated servers.";
+		private const string Description = "Static fields should not have client-specific or mutable data, otherwise they will not be safe in undifferentiated servers.";
+		internal const string MessageFormat = "The static field or property {0} is unsafe because {1} is mutable.";
+
+		private static readonly DiagnosticDescriptor Rule = new DiagnosticDescriptor(
+			DiagnosticId,
+			Title,
+			MessageFormat,
+			Category,
+			DiagnosticSeverity.Error,
+			isEnabledByDefault: true,
+			description: Description
+		);
+
+		public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => ImmutableArray.Create( Rule );
+
+		private readonly MutabilityInspector m_immutabilityInspector = new MutabilityInspector();
+		private readonly Utils m_utils = new Utils();
+
+		public override void Initialize( AnalysisContext context ) {
+			context.RegisterCompilationStartAction( RegisterIfNotTestAssembly );
+		}
+
+		private void RegisterIfNotTestAssembly( CompilationStartAnalysisContext compilation ) {
+			var references = compilation.Compilation.ReferencedAssemblyNames;
+			if( references.Any( r => r.Name.ToUpper().Contains( "NUNIT" ) ) ) {
+				// Compilation is a test assembly, skip
+				return;
+			}
+
+			compilation.RegisterSyntaxNodeAction( AnalyzeField, SyntaxKind.FieldDeclaration );
+			compilation.RegisterSyntaxNodeAction( AnalyzeProperty, SyntaxKind.PropertyDeclaration );
+		}
+
+		private void AnalyzeField( SyntaxNodeAnalysisContext context ) {
+			if( m_utils.IsGeneratedCodefile( context.Node.SyntaxTree.FilePath ) ) {
+				// skip code-gen'd files; they have been hand-inspected to be safe
+				return;
+			}
+
+			var root = context.Node as FieldDeclarationSyntax;
+			if( root == null ) {
+				return;
+			}
+
+			if( !root.Modifiers.Any( SyntaxKind.StaticKeyword ) ) {
+				// ignore non-static
+				return;
+			}
+
+			foreach( var variable in root.Declaration.Variables ) {
+				var symbol = context.SemanticModel.GetDeclaredSymbol( variable ) as IFieldSymbol;
+
+				if( symbol == null ) {
+					continue;
+				}
+
+#pragma warning disable CS0618 // Type or member is obsolete
+				if( symbol.GetAttributes().Any( a => a.AttributeClass.MetadataName == "Unaudited" ) ) {
+					// anyhing marked unaudited should not break the build, it's temporary
+					return;
+				}
+#pragma warning restore CS0618 // Type or member is obsolete
+
+				if( symbol.GetAttributes().Any( a => a.AttributeClass.MetadataName == "Audited" ) ) {
+					// anything marked audited has been explicitly marked as a safe static
+					return;
+				}
+
+				if( m_immutabilityInspector.IsFieldMutable( symbol ) ) {
+					var diagnostic = Diagnostic.Create( Rule, variable.GetLocation(), variable.Identifier.ValueText, "it" );
+					context.ReportDiagnostic( diagnostic );
+					return;
+				}
+
+				InspectType(
+					context,
+					symbol.Type,
+					variable.Initializer?.Value,
+					variable.GetLocation(),
+					variable.Identifier.ValueText
+				);
+			}
+		}
+
+		private void AnalyzeProperty( SyntaxNodeAnalysisContext context ) {
+			if( m_utils.IsGeneratedCodefile( context.Node.SyntaxTree.FilePath ) ) {
+				// skip code-gen'd files; they have been hand-inspected to be safe
+				return;
+			}
+
+			var root = context.Node as PropertyDeclarationSyntax;
+			if( root == null ) {
+				return;
+			}
+
+			if( !root.Modifiers.Any( SyntaxKind.StaticKeyword ) ) {
+				// ignore non-static
+				return;
+			}
+
+			var prop = context.SemanticModel.GetDeclaredSymbol( root );
+			if( prop == null ) {
+				return;
+			}
+
+#pragma warning disable CS0618 // Type or member is obsolete
+			if( prop.GetAttributes().Any( a => a.AttributeClass.MetadataName == "Unaudited" ) ) {
+				// anyhing marked unaudited should not break the build, it's temporary
+				return;
+			}
+#pragma warning restore CS0618 // Type or member is obsolete
+
+			if( prop.GetAttributes().Any( a => a.AttributeClass.MetadataName == "Audited" ) ) {
+				// anything marked audited has been explicitly marked as a safe static
+				return;
+			}
+
+			if( m_immutabilityInspector.IsPropertyMutable( prop ) ) {
+				var diagnostic = Diagnostic.Create( Rule, root.GetLocation(), prop.Name, "it" );
+				context.ReportDiagnostic( diagnostic );
+				return;
+			}
+			if( root.IsPropertyGetterImplemented() ) {
+				// property has getter with body; it is either backed by a field, or is a static function; ignore
+				return;
+			}
+
+			InspectType( context, prop.Type, root.Initializer?.Value, root.GetLocation(), prop.Name );
+		}
+
+		private void InspectType(
+			SyntaxNodeAnalysisContext context,
+			ITypeSymbol type,
+			ExpressionSyntax exp,
+			Location location,
+			string fieldOrPropName
+		) {
+			if( m_immutabilityInspector.IsTypeMarkedImmutable( type ) ) {
+				// if the type is marked immutable, skip checking it, to avoid reporting a diagnostic for each usage of non-immutable types that are marked immutable (another analyzer catches this already)
+				return;
+			}
+
+			// try to use the concrete type if we have it (via a constructor)
+			var flags = MutabilityInspectionFlags.Default;
+			var constructedType = GetConstructedType( context, exp );
+			if( constructedType != null ) {
+				type = constructedType;
+				flags |= MutabilityInspectionFlags.AllowUnsealed;
+			}
+
+			if( m_immutabilityInspector.IsTypeMutable( type, flags ) ) {
+				var diagnostic = Diagnostic.Create( Rule, location, fieldOrPropName, type.GetFullTypeNameWithGenericArguments() );
+				context.ReportDiagnostic( diagnostic );
+			}
+		}
+
+		private ITypeSymbol GetConstructedType(
+			SyntaxNodeAnalysisContext context,
+			ExpressionSyntax exp
+		) {
+
+			if( exp == null ) {
+				return null;
+			}
+
+			var initializerSymbol = context.SemanticModel.GetSymbolInfo( exp ).Symbol as IMethodSymbol;
+			if( initializerSymbol == null ) {
+				return null;
+			}
+
+			if( initializerSymbol.MethodKind != MethodKind.Constructor ) {
+				return null;
+			}
+
+			return initializerSymbol.ContainingType;
+		}
+
+	}
+}
