--- tests/D2L.CodeStyle.Analysis.UnitTests/Properties/AssemblyInfo.cs
@@ -1,14 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-using NUnit.Framework;
-
-[assembly: AssemblyTitle("D2L.CodeStyle.Analysis.UnitTests")]
-
-[assembly: ComVisible(false)]
-
-[assembly: Guid("e5e9d471-bb72-4f77-88f8-3b4f77a7d23b")]
-
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
-
-[assembly: Category("Unit")]
