+++ tests/D2L.CodeStyle.Analyzers.Test/Common/SyntaxNodeExtensionTests.cs
@@ -1,9 +1,9 @@
 ﻿using Microsoft.CodeAnalysis.CSharp.Syntax;
 using NUnit.Framework;
 using System.Linq;
-using static D2L.CodeStyle.Analysis.RoslynSymbolFactory;
+using static D2L.CodeStyle.Analyzers.Common.RoslynSymbolFactory;
 
-namespace D2L.CodeStyle.Analysis {
+namespace D2L.CodeStyle.Analyzers.Common {
 
     [TestFixture]
     internal sealed class SyntaxNodeExtensionTests {
